[feat] ci/cd 파이프라인 개선 (#55)
* #51 fix: cicd 개선

- docker 추가

* #51 fix: cicd 테스트 브랜치 추가

* #10 feat: getFile -> getInputStream으로 변경

* #10 feat: 엘라스틱 서치 코드 제거 및 REST API 검색 구현

* #10 feat: 엘라스틱 서치 컨테이너 제거

* #10 chore: 엘라스틱 서치 의존성 제거

* #51 fix: 경로 변경
