+++ .github/workflows/cicd.yml
@@ -2,7 +2,7 @@ name: ci-cd
 
 on:
   push:
-    branches: [ be/docs/#7-cicd, dev ]
+    branches: [ be/infra/#51-cicd, dev ]
 
 permissions:
   contents: read
@@ -52,6 +52,11 @@ jobs:
           touch ./application-secret.yml
           echo "${{ secrets.APPLICATION_SECRET }}" > ./application-secret.yml
         shell: bash
+      # env 파일 생성
+      - name: make .env file
+        run: |
+          touch .env
+          echo "${{ secrets.ENV }}" > ./.env
       # gradlew 실행을 위해서 실행 권한을 부여
       - name: Grant execute permission for gradlew
         run: chmod +x ./gradlew
