+++ fineAnts_app/Dockerfile
@@ -1,0 +1,6 @@
+FROM openjdk:11
+WORKDIR /app
+COPY . .
+ARG JAR_FILE=./build/libs/*.jar
+COPY ${JAR_FILE} app.jar
+ENTRYPOINT ["java","-Dspring.profiles.active=dev","-jar","app.jar"]
