+++ fineAnts_app/build.gradle
@@ -36,8 +36,6 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-webflux'
-    // elasticSearch
-    implementation 'org.springframework.data:spring-data-elasticsearch'
 
     // websocket
     implementation 'org.springframework.boot:spring-boot-starter-websocket'
