+++ fineAnts_app/docker-compose-dev.yml
@@ -1,0 +1,24 @@
+version: "3.8"
+services:
+  app:
+    container_name: fineAnts_app
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - 8080:8080
+    environment:
+      - TZ=Asia/Seoul
+    depends_on:
+      - redis
+    networks:
+      - backend_net
+  redis:
+    container_name: fineAnts_redis
+    image: redis:latest
+    ports:
+      - "6379:6379"
+    networks:
+      - backend_net
+networks:
+  backend_net:
