+++ fineAnts_app/docker-compose.yml
@@ -1,18 +1,5 @@
 version: "3.8"
 services:
-  elasticsearch:
-    container_name: fineAnts_es
-    image: docker.elastic.co/elasticsearch/elasticsearch:8.10.4
-    platform: linux/x86_64
-    ports:
-      - "9200:9200"
-    environment:
-      - discovery.type=single-node
-      - xpack.security.enabled=${ELASTIC_SECURITY}
-      - ELASTICSEARCH_USERNAME=${ELASTICSEARCH_USERNAME}
-      - ELASTIC_PASSWORD=${ELASTIC_PASSWORD}
-    networks:
-      - backend_net
   db:
     container_name: fineAnts_db
     image: mysql:latest
