--- fineAnts_app/src/main/java/codesquad/fineants/elasticsearch/configuration/ElasticSearchConfig.java
@@ -1,56 +1,0 @@
-package codesquad.fineants.elasticsearch.configuration;
-
-import org.apache.http.Header;
-import org.apache.http.HttpHeaders;
-import org.apache.http.HttpHost;
-import org.apache.http.auth.AuthScope;
-import org.apache.http.auth.UsernamePasswordCredentials;
-import org.apache.http.impl.client.BasicCredentialsProvider;
-import org.apache.http.message.BasicHeader;
-import org.elasticsearch.client.RestClient;
-import org.elasticsearch.client.RestClientBuilder;
-import org.elasticsearch.client.RestHighLevelClient;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories;
-
-@Configuration
-@EnableElasticsearchRepositories(basePackages = "codesquad.fineants.elasticsearch.repository")
-@ComponentScan(basePackages = {"codesquad.fineants.elasticsearch"})
-public class ElasticSearchConfig {
-
-	@Value("${elasticsearch.host}")
-	private String host;
-
-	@Value("${elasticsearch.port}")
-	private int port;
-
-	@Value("${elasticsearch.protocol}")
-	private String protocol;
-
-	@Value("${elasticsearch.username}")
-	private String userName;
-
-	@Value("${elasticsearch.password}")
-	private String password;
-
-	@Bean(destroyMethod = "close")
-	public RestHighLevelClient restClient() {
-		final BasicCredentialsProvider credentialsProvider = new BasicCredentialsProvider();
-		credentialsProvider.setCredentials(AuthScope.ANY, new UsernamePasswordCredentials(userName, password));
-
-		RestClientBuilder builder = RestClient.builder(new HttpHost(host, port, protocol))
-			.setHttpClientConfigCallback(
-				httpClientBuilder -> httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider))
-			.setDefaultHeaders(compatibilityHeaders());
-		return new RestHighLevelClient(builder);
-	}
-
-	private Header[] compatibilityHeaders() {
-		return new Header[] {
-			new BasicHeader(HttpHeaders.ACCEPT, "application/vnd.elasticsearch+json;compatible-with=7"),
-			new BasicHeader(HttpHeaders.CONTENT_TYPE, "application/vnd.elasticsearch+json;compatible-with=7")};
-	}
-}
