--- fineAnts_app/src/main/java/codesquad/fineants/elasticsearch/helper/Indices.java
@@ -1,5 +1,0 @@
-package codesquad.fineants.elasticsearch.helper;
-
-public final class Indices {
-	public static final String STOCK_INDEX = "stocksearch";
-}
