--- fineAnts_app/src/main/java/codesquad/fineants/elasticsearch/repository/StockSearchRepository.java
@@ -1,8 +1,0 @@
-package codesquad.fineants.elasticsearch.repository;
-
-import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
-
-import codesquad.fineants.elasticsearch.document.StockSearch;
-
-public interface StockSearchRepository extends ElasticsearchRepository<StockSearch, String> {
-}
