--- fineAnts_app/src/main/java/codesquad/fineants/elasticsearch/search/SearchRequestDTO.java
@@ -1,17 +1,0 @@
-package codesquad.fineants.elasticsearch.search;
-
-import java.util.List;
-
-import org.elasticsearch.search.sort.SortOrder;
-
-import lombok.Getter;
-import lombok.ToString;
-
-@Getter
-@ToString
-public class SearchRequestDTO {
-	private List<String> fields;
-	private String searchTerm;
-	private String sortBy;
-	private SortOrder order;
-}
