+++ fineAnts_app/src/main/java/codesquad/fineants/spring/api/stock/StockRestController.java
@@ -1,0 +1,27 @@
+package codesquad.fineants.spring.api.stock;
+
+import java.util.List;
+
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import codesquad.fineants.spring.api.response.ApiResponse;
+import codesquad.fineants.spring.api.stock.request.StockSearchRequest;
+import codesquad.fineants.spring.api.stock.response.StockSearchItem;
+import codesquad.fineants.spring.api.success.code.StockSuccessCode;
+import lombok.RequiredArgsConstructor;
+
+@RequestMapping("/api/stocks")
+@RequiredArgsConstructor
+@RestController
+public class StockRestController {
+
+	private final StockService stockService;
+
+	@PostMapping("/search")
+	public ApiResponse<List<StockSearchItem>> search(@RequestBody final StockSearchRequest request) {
+		return ApiResponse.success(StockSuccessCode.OK_SEARCH_STOCKS, stockService.search(request));
+	}
+}
