+++ fineAnts_app/src/main/java/codesquad/fineants/spring/api/stock/request/StockSearchRequest.java
@@ -1,0 +1,8 @@
+package codesquad.fineants.spring.api.stock.request;
+
+import lombok.Getter;
+
+@Getter
+public class StockSearchRequest {
+	private String searchTerm;
+}
