--- fineAnts_app/src/main/resources/static/es-settings.json
@@ -1,4 +1,0 @@
-{
-  "index": {
-  }
-}
